pyproject.toml
[tool.poetry]
name = "rag-system"
version = "1.0.0"
description = "RAG 智能问答系统 - 基于 LangChain + Chroma + MiniMax 的多用户智能文档问答系统"
authors = ["Jonas"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^1.0.0"
langchain-anthropic = "^1.0.0"
langchain-core = "^1.0.0"
anthropic = "^0.69.0"
langchain-community = "^0.4.1"
filelock = "^3.20.0"
chromadb = "^1.3.4"
langchain-chroma = "^1.0.0"
pypdf = "^6.2.0"
sentence-transformers = "^5.1.2"
langchain-huggingface = "^1.0.1"
tenacity = "^9.1.2"
streamlit = "^1.51.0"
bcrypt = "^5.0.0"
pyjwt = "^2.10.1"
python-docx = "^1.2.0"
python-dotenv = "^1.2.1"
python-multipart = "^0.0.20"
streamlit-cookies-controller = "^0.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

